fix: align desktop clerk bridge config
